Fix for bug introduced for database loading
When loading a database with allow_unfound=True and without
database inputs and targets, the code would fail due to the
inability to process the database.var_list in this case.

Now, in this case, the check is skipped, as in prior versions,
but now information about skipping the check is printed.